apps/api/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
typer
pytest
httpx
tqdm
pyyaml
ruff
mypy
